requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
numpy<2.0.0
Pillow==9.5.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
python-magic-bin==0.4.14
pydantic==2.5.0
pydantic-settings==2.1.0
sentencepiece==0.1.99
accelerate==0.25.0
easyocr==1.7.1
opencv-python-headless==4.8.1.78